update API work ; MVC is still not correct
