+++ TravelerAPI/TravelerAPI/Currency/Model/CurrencyService.swift
@@ -7,43 +7,38 @@
 
 import Foundation
 
-class CurrencyService {
+final class CurrencyService {
+    
     // MARK: Properties
     private let currencyUrl = "https://api.apilayer.com/fixer/symbols"
     private let apiKey = "RsMNr3axWpCqqqI2viaMbRnLmayHmFEd"
-    
-    
-    // MARK: Json Parsing and APIRest
     private var task:URLSessionDataTask?
     
-    func getCurrency(callBack: @escaping (Bool, [String]?) -> Void) {
+    // MARK: Actions
+    func getCurrency(completionHandler: @escaping (Bool, Currency?) -> Void) {
         guard let url = URL(string: currencyUrl) else { return }
+        
         var request = URLRequest(url:url)
         request.httpMethod = "GET"
         request.addValue(apiKey, forHTTPHeaderField: "apikey")
         
-      let session = URLSession(configuration: .default)
+        let session = URLSession(configuration: .default)
         let task = session.dataTask(with: request) { data, response, error in
             DispatchQueue.main.async {
                 guard let data = data, error == nil else {
-                    callBack(false, nil)
+                    completionHandler(false, nil)
                     return
                 }
                 guard let response = response as? HTTPURLResponse, response.statusCode == 200 else {
-                    callBack(false, nil)
+                    completionHandler(false, nil)
                     return
                 }
                 guard let responseJSON = try? JSONDecoder().decode(Currency.self, from: data) else {
-                    callBack(false, nil)
+                    completionHandler(false, nil)
                     return
                 }
-                
-                let symbolsJSON = responseJSON.symbols.map({(symbol, name) in symbol})
-                //let nameSymbolsJSON = responseJSON.symbols.map({(_, name) in name})
-                
-                callBack(true, symbolsJSON)
+                completionHandler(true, responseJSON)
             }
-                    
         }
         task.resume()
     }
