+++ TravelerAPI/TravelerAPI/Supported Files/SceneDelegate.swift
@@ -15,7 +15,7 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
             
             guard let windowScene = (scene as? UIWindowScene) else { return }
             window = UIWindow(frame: UIScreen.main.bounds)
-            let home = TravelTabBarViewController()
+            let home = TravelTabBarController()
             home.selectedIndex = 2
             self.window?.rootViewController = home
             window?.makeKeyAndVisible()
